issue #20 - soap types split into file
